change variable names and put the activities in a separate file
