+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.HumanLearning">
-        <activity android:name=".DisplayImageActivity"></activity>
+        <activity android:name=".DisplayDatasetActivity.DisplayImageActivity" />
         <activity android:name=".MainActivity" />
-        <activity android:name=".DisplayDatasetActivity">
+        <activity android:name=".DisplayDatasetActivity.DisplayDatasetActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
