+++ app/src/main/res/layout/activity_display_image.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DisplayImageActivity">
+    tools:context=".DisplayDatasetActivity.DisplayImageActivity">
 
     <ImageView
-        android:id="@+id/viewImage"
+        android:id="@+id/display_image_viewImage"
         android:layout_width="400dp"
         android:layout_height="400dp"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -16,13 +16,13 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/viewCategory"
+        android:id="@+id/display_image_viewCategory"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textSize="34sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/viewImage" />
+        app:layout_constraintTop_toBottomOf="@+id/display_image_viewImage" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
