+++ app/src/main/res/layout/image_and_category_item.xml
@@ -13,7 +13,7 @@
         android:layout_height="wrap_content">
 
         <ImageView
-            android:id="@+id/imageView"
+            android:id="@+id/image_and_category_item_imageView"
             android:layout_width="wrap_content"
             android:layout_height="150dp"
             android:layout_centerHorizontal="true"
@@ -21,11 +21,10 @@
             android:src="@mipmap/ic_launcher" />
 
         <TextView
-            android:id="@+id/imageCategory"
+            android:id="@+id/image_and_category_item_imageCategory"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_below="@id/imageView"
-            android:layout_centerHorizontal="true"
-            android:text="@string/category" />
+            android:layout_below="@id/image_and_category_item_imageView"
+            android:layout_centerHorizontal="true"/>
     </RelativeLayout>
 </androidx.cardview.widget.CardView>
