+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">My Applicationn</string>
     <string name="category">Category</string>
+    <string name="image_and_category_item_category">image_and_category_item_category</string>
 </resources>
